Pull out constants and utility functions.
